# Conflicts:
#	code/go/0chain.net/smartcontract/dbs/event/delegate_pool.go

+++ code/go/0chain.net/smartcontract/storagesc/validator_eventdb.go
@@ -103,7 +103,7 @@ func emitAddOrOverwriteValidator(
 		OffersTotal:  sp.TotalOffers,
 		UnstakeTotal: sp.TotalUnStake,
 		Reward:       sp.Reward,
-		TotalStake:   int64(sp.stake()),
+		TotalStake:   sp.stake(),
 	})
 	if err != nil {
 		return fmt.Errorf("marshalling validator: %v", err)
